re-add tests and ensure they work
These were the following changes I made:

 1. estimate_r changed to estimate_R
 2. plot = FALSE was no longer a parameter
 3. The new output was a class "estimate_R", but the saved results weren't

Example 2, however fails as I mentioned earlier and it appears to be due
to the sampling bug that was fixed in R 3.6. Since this is a trivial
sampling, I'm going to hard-code the values to avoid having to set or
reset the default sampler.